Initial commit of TRView SWT based application.
This application serves as a client side of TRV. It is used to connect to server side of TRV and receive JUnit data (client package). Received data are parsed and displayed to user in swt window (view package).

Note: class files may be removed later. They are here because folder was copied directly from workspace.