Remove RedDeer dependency from InRunJUnit plugin + formatting code style.
Note: Initialy, RedDeer was added because of parsing results. Later, when everything was handled by JUnit plugin it was there because of logging. Now it is useless to import whole RedDeer just because of logging. Furthermore, it may cause problems while installing with other plugins. Error logs are now logged using System.err stream.